RoleAccessChecher funcionando en planestudio, planmateria y programa, se agregan a actionroles

+++ controllers/PlanmateriaController.php
@@ -8,8 +8,11 @@ use app\models\Planestudio;
 use app\models\Materia;
 use app\models\PlanemateriaSearch;
 use yii\web\Controller;
+use yii\filters\AccessControl;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use app\commands\RoleAccessChecker;
+use app\controllers\ErrorController;
 
 /**
  * PlanmateriaController implements the CRUD actions for Planmateria model.
@@ -22,6 +25,27 @@ class PlanmateriaController extends Controller
     public function behaviors()
     {
         return [
+              'access' => [
+                 'class' => AccessControl::className(),
+                 'only' => ['index', 'view', 'update', 'create', 'delete',],
+                 'rules' => [
+                     [
+                         'allow' => true,
+                         'actions' => ['',],
+                         'roles' => ['?'],
+                     ],
+                     [
+                         'allow' => true,
+                         'actions' => ['index', 'view', 'update', 'create', 'delete',],
+                         'roles' => ['@'],
+                     ],
+                     [
+                         'allow' => false,
+                         'actions' => ['',],
+                         'roles' => ['@'],
+                     ],
+                 ],
+             ],
             'verbs' => [
                 'class' => VerbFilter::className(),
                 'actions' => [
@@ -35,15 +59,17 @@ class PlanmateriaController extends Controller
      * Lists all Planmateria models.
      * @return mixed
      */
-    public function actionIndex()
-    {
-        $searchModel = new PlanemateriaSearch();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+    public function actionIndex(){
+		$msg='';
+		if (RoleAccessChecker::actionIsAsignSector('planmateria/index')) {
+			$searchModel = new PlanemateriaSearch();
+			$dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+			return $this->render('index', [
+				'searchModel' => $searchModel,
+				'dataProvider' => $dataProvider,
+			]);
+        }else return $this->redirect(['error/error',["msg" => $msg ]]);
     }
 
     /**
@@ -51,11 +77,13 @@ class PlanmateriaController extends Controller
      * @param integer $id
      * @return mixed
      */
-    public function actionView($id)
-    {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+    public function actionView($id){
+		$msg='';
+		if (RoleAccessChecker::actionIsAsignSector('planmateria/view')) {
+			return $this->render('view', [
+				'model' => $this->findModel($id),
+			]);
+        }else return $this->redirect(['error/error',["msg" => $msg ]]);
     }
 
     /**
@@ -63,20 +91,22 @@ class PlanmateriaController extends Controller
      * If creation is successful, the browser will be redirected to the 'view' page.
      * @return mixed
      */
-    public function actionCreate()
-    {
-        $model = new Planmateria();
-        $subModel= new Planestudio();
-        $subModel2= new Materia();
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['index', 'id' => $model->planmateria_id]);
-        } else {
-            return $this->render('create', [
-                'model' => $model,
-                'subModel' => $subModel,
-                'subModel2' => $subModel2,
-            ]);
-        }
+    public function actionCreate(){
+		$msg='';
+		if (RoleAccessChecker::actionIsAsignSector('planmateria/create')) {
+			$model = new Planmateria();
+			$subModel= new Planestudio();
+			$subModel2= new Materia();
+			if ($model->load(Yii::$app->request->post()) && $model->save()) {
+				return $this->redirect(['index', 'id' => $model->planmateria_id]);
+			} else {
+				return $this->render('create', [
+					'model' => $model,
+					'subModel' => $subModel,
+					'subModel2' => $subModel2,
+				]);
+			}
+        }else return $this->redirect(['error/error',["msg" => $msg ]]);
     }
 
     /**
@@ -85,22 +115,24 @@ class PlanmateriaController extends Controller
      * @param integer $id
      * @return mixed
      */
-    public function actionUpdate($id)
-    {
-        $model = $this->findModel($id);
-        $subModel= new Planestudio();
-        $subModel2= new Materia();
+    public function actionUpdate($id){
+		$msg='';
+		if (RoleAccessChecker::actionIsAsignSector('planmateria/update')) {
+			$model = $this->findModel($id);
+			$subModel= new Planestudio();
+			$subModel2= new Materia();
 
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['index', 'id' => $model->planmateria_id]);
-        } else {
-            return $this->render('update', [
-                'model' => $model,
-                'subModel' => $subModel,
-                'subModel2' => $subModel2,
-            ]);
-        }
+			if ($model->load(Yii::$app->request->post()) && $model->save()) {
+				return $this->redirect(['index', 'id' => $model->planmateria_id]);
+			} else {
+				return $this->render('update', [
+					'model' => $model,
+					'subModel' => $subModel,
+					'subModel2' => $subModel2,
+				]);
+			}
+        }else return $this->redirect(['error/error',["msg" => $msg ]]);
     }
 
     /**
@@ -109,11 +141,12 @@ class PlanmateriaController extends Controller
      * @param integer $id
      * @return mixed
      */
-    public function actionDelete($id)
-    {
-        $this->findModel($id)->delete();
-
-        return $this->redirect(['index']);
+    public function actionDelete($id){
+		$msg='';
+		if (RoleAccessChecker::actionIsAsignSector('planmateria/delete')) {
+			$this->findModel($id)->delete();
+			return $this->redirect(['index']);
+        }else return $this->redirect(['error/error',["msg" => $msg ]]);
     }
 
     /**
